% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allBiplotsFunction.R
\name{allBiplots}
\alias{allBiplots}
\title{Save all Biplots}
\usage{
allBiplots(df, mColors, showSources = F, onlySources = F)
}
\arguments{
\item{df}{dataframe from the elemSource function}

\item{mColors}{optional list of colors to use in plot}

\item{showSources}{logical- display the sources along with the sourced artifacts}

\item{onlySources}{logical- only run the sources}
}
\value{
None
}
\description{
Function to run and save all possible combinations of biplots
}
\examples{
allBiplots(df)

}
