% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemSource.R, R/selectData.R
\name{elemSource}
\alias{elemSource}
\alias{elemSource}
\title{Function to source data}
\usage{
elemSource(df, saveResults = T, prob = 0.9)

elemSource(df, saveResults = T, prob = 0.9)
}
\arguments{
\item{df}{dataframe to source}

\item{saveResults}{logical-save results}

\item{prob}{= probability to accept sourcing}

\item{df}{dataframe to source}

\item{saveResults}{logical-save results}

\item{prob}{= probability to accept sourcing}
}
\value{
None

None
}
\description{
Runs function to source data

Runs function to source data
}
\examples{
elemSource(df)

elemSource(df)

}
