% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Save PCA Plots}
\usage{
plotPCA(df, mColors, showSources = F, onlySources = F)
}
\arguments{
\item{df}{dataframe to source}

\item{mColors}{optional list of colors to use in plot}

\item{showSources}{logical- display the sources along with the sourced artifacts}

\item{onlySources}{logical- only run the sources}
}
\value{
None
}
\description{
Function to create and save PCA plots
}
\examples{
elemSource(df)

}
