% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBiplotApp.R
\name{runBiplotApp}
\alias{runBiplotApp}
\title{Function to source data}
\usage{
runBiplotApp()
}
\arguments{
\item{None}{dataframe to source}
}
\value{
None
}
\description{
Runs function to source data
}
\examples{
runBiplotApp(df)

}
