% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTriplotApp.R
\name{runTriplotApp}
\alias{runTriplotApp}
\title{Run Triplot Shiny App}
\usage{
runTriplotApp(df)
}
\arguments{
\item{df}{Dataframe that has been sourced using the elemSource() function}
}
\value{
None
}
\description{
Function to run Triplot shiny app
}
\examples{
runTriplotApp())

}
