% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlots.R
\name{scatterPlots}
\alias{scatterPlots}
\title{Create all Biplots}
\usage{
scatterPlots(df = "dataframe", groupName = NULL, mColors = NULL)
}
\arguments{
\item{df}{dataframe to source}

\item{groupName}{Name of column to group by}

\item{mColors}{optional list of colors to use in plot}
}
\value{
None
}
\description{
Function to create all possible biplots
}
\examples{
scatterPlots(df)

}
