% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectData.R
\name{selectData}
\alias{selectData}
\title{Function to source data}
\usage{
selectData()
}
\arguments{
\item{df}{dataframe to source}

\item{saveResults}{logical-save results}

\item{prob}{= probability to accept sourcing}
}
\value{
None
}
\description{
Runs function to source data
}
\examples{
elemSource(df)

}
