% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBiplotApp.R
\name{runBiplotApp}
\alias{runBiplotApp}
\title{Run Biplot Shiny App}
\usage{
runBiplotApp(df)
}
\arguments{
\item{df}{Dataframe that has been sourced using the elemSource() function}
}
\value{
None
}
\description{
Function to run Biplot shiny app
}
\examples{
runBiplotApp(df)

}
