% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTriplotApp.R
\name{runTriplotApp}
\alias{runTriplotApp}
\title{Function to source data}
\usage{
runTriplotApp()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
Runs function to source data
}
\examples{
runTriplotApp())

}
