% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTriplotApp.R
\name{runTriplotApp}
\alias{runTriplotApp}
\title{Run Triplot Shiny App}
\usage{
runTriplotApp()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
Function to run Triplot shiny app
}
\examples{
runTriplotApp())

}
